requirements.txt
Flask==2.2.2
SQLAlchemy==1.4.31
mysql-connector-python==8.0.28
requests==2.27.1
ollama==0.1.0
streamlit==1.10.0
chromadb==0.4.18
sentence-transformers==2.2.2
torch==2.1.0
numpy>=1.24.0